% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oz.R
\name{Oz}
\alias{Oz}
\title{Most probable state (similar to Viterbi)}
\usage{
Oz(nhmmobj, outfile = NULL)
}
\arguments{
\item{nhmmobj}{an object created from the NHMM function}

\item{outfile}{a directory to put the .png plot}
}
\value{
zbest the most probable sequence from all iterations

output: a plot of a histogram of the distribution of 
the most probable state sequence. If the number of states in 
the histogram is less than K, it probably means you should probably
re-run the model with smaller K as some of the states have disappeared.
}
\description{
\code{Oz} calculates the most probable state per time step (Viterbi like) with values from 1,...,K.
 The histogram of this sequence is displayed in the GUI output.   
 If there are ties for a given day then the lowest number state is chosen.
}
\examples{
#Oz(my.nhmm) 
}
